Factor out src/Prelude.hs
